improve tenant manager client, disable fail on unknown properties
