+++ app/src/main/java/io/apicurio/registry/mt/TenantMetadataService.java
@@ -16,8 +16,8 @@
 
 package io.apicurio.registry.mt;
 
-import io.apicurio.multitenant.api.beans.TenantStatusValue;
 import io.apicurio.multitenant.api.datamodel.RegistryTenant;
+import io.apicurio.multitenant.api.datamodel.TenantStatusValue;
 import io.apicurio.multitenant.api.datamodel.UpdateRegistryTenantRequest;
 import io.apicurio.multitenant.client.TenantManagerClient;
 import io.apicurio.multitenant.client.exception.RegistryTenantNotFoundException;
