+++ app/src/main/java/io/apicurio/registry/mt/TenantReaper.java
@@ -16,11 +16,11 @@
 
 package io.apicurio.registry.mt;
 
-import io.apicurio.multitenant.api.beans.RegistryTenantList;
-import io.apicurio.multitenant.api.beans.SortBy;
-import io.apicurio.multitenant.api.beans.SortOrder;
-import io.apicurio.multitenant.api.beans.TenantStatusValue;
 import io.apicurio.multitenant.api.datamodel.RegistryTenant;
+import io.apicurio.multitenant.api.datamodel.RegistryTenantList;
+import io.apicurio.multitenant.api.datamodel.SortBy;
+import io.apicurio.multitenant.api.datamodel.SortOrder;
+import io.apicurio.multitenant.api.datamodel.TenantStatusValue;
 import io.apicurio.multitenant.client.TenantManagerClient;
 import io.apicurio.registry.storage.RegistryStorage;
 import io.apicurio.registry.types.Current;
