+++ app/src/test/java/io/apicurio/registry/mt/limits/MultitenancyLimitsTest.java
@@ -25,13 +25,13 @@ import java.util.function.Supplier;
 
 import javax.inject.Inject;
 
-import io.apicurio.multitenant.api.beans.TenantStatusValue;
 import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.opentest4j.TestAbortedException;
 
 import io.apicurio.multitenant.api.datamodel.RegistryTenant;
+import io.apicurio.multitenant.api.datamodel.TenantStatusValue;
 import io.apicurio.registry.mt.MockTenantMetadataService;
 import io.apicurio.registry.rest.client.RegistryClient;
 import io.apicurio.registry.rest.client.RegistryClientFactory;
