+++ multitenancy/tenant-manager-api/src/main/java/io/apicurio/multitenant/storage/dto/RegistryTenantDto.java
@@ -30,8 +30,8 @@ import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
-import io.apicurio.multitenant.api.beans.TenantStatusValue;
 import io.apicurio.multitenant.api.datamodel.RegistryTenant;
+import io.apicurio.multitenant.api.datamodel.TenantStatusValue;
 
 /**
  * @author Fabian Martinez
