+++ multitenancy/tenant-manager-datamodel/src/main/java/io/apicurio/multitenant/api/datamodel/RegistryTenant.java
@@ -10,7 +10,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import io.apicurio.multitenant.api.beans.TenantStatusValue;
 
 
 /**
