+++ multitenancy/tenant-manager-datamodel/src/main/java/io/apicurio/multitenant/api/datamodel/RegistryTenantList.java
@@ -1,5 +1,5 @@
 
-package io.apicurio.multitenant.api.beans;
+package io.apicurio.multitenant.api.datamodel;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -8,7 +8,6 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import io.apicurio.multitenant.api.datamodel.RegistryTenant;
 
 
 /**
