+++ multitenancy/tenant-manager-datamodel/src/main/java/io/apicurio/multitenant/api/datamodel/TenantStatusValue.java
@@ -1,5 +1,5 @@
 
-package io.apicurio.multitenant.api.beans;
+package io.apicurio.multitenant.api.datamodel;
 
 import java.util.HashMap;
 import java.util.Map;
